[Indexer] Create `objects_version` table.
## Description

This table maps an object's ID and version to a checkpoint sequence
number, in a table partitioned by the first byte of the object ID.
This speeds up look ups into `objects_history` by offering a path for
a first look-up to the correct partition in that table for a given
object's ID and version.

This PR introduces the table, and the logic to populate it in the
indexer.

## Test plan

```
sui$ cargo nextest run -p sui-indexer
sui$ cargo nextest run -p sui-graphql-rpc
sui$ cargo nextest run -p sui-graphql-e2e-tests --features pg_integration
```